psh: code split into features
JIRA: BES-105
