+++ core/psh/cat-empty.c
@@ -1,0 +1,28 @@
+/*
+ * Phoenix-RTOS
+ *
+ * cat - concatenate file(s) to standard output (empty implementation)
+ *
+ * Copyright 2021 Phoenix Systems
+ * Author: Mateusz Niewiadomski
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#include <errno.h>
+#include <stdio.h>
+
+
+void psh_catinfo(void)
+{
+	return;
+}
+
+
+int psh_cat(int argc, char **argv)
+{
+	fprintf(stderr,"Command not supported!\n");
+	return -ENOTSUP;
+}
