+++ core/psh/help-empty.c
@@ -1,0 +1,28 @@
+/*
+ * Phoenix-RTOS
+ *
+ * Phoenix-RTOS SHell
+ *
+ * Copyright 2017, 2018, 2020, 2021 Phoenix Systems
+ * Author: Pawel Pisarczyk, Jan Sikorski, Lukasz Kosinski, Mateusz Niewiadomski
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#include <errno.h>
+#include <stdio.h>
+
+
+void psh_helpinfo(void)
+{
+	return;
+}
+
+
+void psh_help(void)
+{
+	fprintf(stderr,"Command not supported!\n");
+	return -ENOTSUP;
+}
