+++ core/psh/ls-empty.c
@@ -1,0 +1,28 @@
+/*
+ * Phoenix-RTOS
+ *
+ * fs - lists files and directories, based on GNU implementation (empty implementation)
+ *
+ * Copyright 2021 Phoenix Systems
+ * Author: Mateusz Niewiadomski
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#include <errno.h>
+#include <stdio.h>
+
+
+void psh_lsinfo(void)
+{
+	return;
+}
+
+
+int psh_ls(int argc, char **argv)
+{
+	fprintf(stderr,"Command not supported!\n");
+	return -ENOTSUP;
+}
