+++ core/psh/mkdir-empty.c
@@ -1,0 +1,28 @@
+/*
+ * Phoenix-RTOS
+ *
+ *  mkdir - creates directory (empty implementation)
+ *
+ * Copyright 2021 Phoenix Systems
+ * Author: Mateusz Niewiadomski
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#include <errno.h>
+#include <stdio.h>
+
+
+void psh_mkdirinfo(void)
+{
+	return;
+}
+
+
+int psh_mkdir(int argc, char **argv)
+{
+	fprintf(stderr,"Command not supported!\n");
+	return -ENOTSUP;
+}
