+++ core/psh/reboot.c
@@ -3,8 +3,8 @@
  *
  * reboot - restart the machine
  *
- * Copyright 2017, 2018, 2020 Phoenix Systems
- * Author: Pawel Pisarczyk, Jan Sikorski, Lukasz Kosinski
+ * Copyright 2017, 2018, 2020, 2021 Phoenix Systems
+ * Author: Pawel Pisarczyk, Jan Sikorski, Lukasz Kosinski, Mateusz Niewiadomski
  *
  * This file is part of Phoenix-RTOS.
  *
@@ -16,6 +16,12 @@
 #include <unistd.h>
 
 #include <sys/reboot.h>
+
+
+void psh_rebootinfo(void)
+{
+	printf("  reboot  - restarts the machine\n");
+}
 
 
 int psh_reboot(int argc, char **argv)
