+++ core/psh/sysexec-empty.c
@@ -1,0 +1,28 @@
+/*
+ * Phoenix-RTOS
+ *
+ * sysexec - launch program from syspage using given map (empty implementation)
+ *
+ * Copyright 2021 Phoenix Systems
+ * Author: Mateusz Niewiadomski
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#include <errno.h>
+#include <stdio.h>
+
+
+void psh_sysexecinfo(void)
+{
+	return;
+}
+
+
+int psh_sysexec(int argc, char **argv)
+{
+	fprintf(stderr,"Command not supported!\n");
+	return -ENOTSUP;
+}
